# Reduce cognitive complexity of the scope parser